docs/onboarding/requirements.txt
# Core dependencies
astroid>=2.15.0
pylint>=2.17.0
mypy>=1.0.0
flake8>=6.0.0
black>=23.0.0
coverage>=7.0.0

# AST manipulation
ast>=0.0.0  # Built-in, listed for documentation

# Type checking
typing-extensions>=4.5.0

# Logging and utilities
colorama>=0.4.6  # For colored terminal output
rich>=13.0.0     # For enhanced logging and formatting

# Configuration
pyyaml>=6.0.0    # For YAML config support

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
isort>=5.12.0    # For import sorting 